Refactor and reorganize files
